+++ CHANGELOG.rst
@@ -6,12 +6,14 @@ Changelog
 
 Release date: ``2020-xx-xx``
 
+- `NXPY-186 <https://jira.nuxeo.com/browse/NXPY-186>`__: Remove the ``Blob.batch_id`` attribute
 - `NXPY-192 <https://jira.nuxeo.com/browse/NXPY-192>`__: Add support for Python 3.10
 - `NXPY-193 <https://jira.nuxeo.com/browse/NXPY-193>`__: Fix thread-safety in uploads and workflows
 
 Technical changes
 -----------------
 
+- Removed ``Blob.batch_id``. Use ``Blob.batchId`` instead.
 - Removed utils.py::``SwapAttr``
 
 3.1.1