+++ CHANGELOG.rst
@@ -1,13 +1,14 @@
 Changelog
 =========
 
-3.1.2
+4.0.0
 -----
 
-Release date: ``2020-xx-xx``
+Release date: ``2020-12-05``
 
 - `NXPY-186 <https://jira.nuxeo.com/browse/NXPY-186>`__: Remove the ``Blob.batch_id`` attribute
-- `NXPY-191 <https://jira.nuxeo.com/browse/NXPY-191>`__: Fix ``urllib3`` DeprecationWarning in client.py
+- `NXPY-188 <https://jira.nuxeo.com/browse/NXPY-188>`__: Add mimetype tests
+- `NXPY-191 <https://jira.nuxeo.com/browse/NXPY-191>`__: Fix ``urllib3`` DeprecationWarning in ``client.py``
 - `NXPY-192 <https://jira.nuxeo.com/browse/NXPY-192>`__: Add support for Python 3.10
 - `NXPY-193 <https://jira.nuxeo.com/browse/NXPY-193>`__: Fix thread-safety in uploads and workflows